requirements.txt
asgiref==3.10.0
PyJWT==2.10.1
python-monkey-business==1.1.0
sqlparse==0.5.3
tzdata==2025.2
apscheduler==3.11.1
schedule==1.2.2
xlsxwriter==3.2.9
weasyprint==66.0
openpyxl==3.1.2


# Django & DRF
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-filter==24.1

# Excel export
openpyxl==3.1.2
pandas==2.2.1

# PDF export
reportlab==4.1.0

# Encrypted fields (optional)
django-fernet-fields==0.6

# Environment variables
python-dotenv==1.0.1

# Strong date/time utilities
pendulum==3.0.0
